Move categories to separate module
